# contracts/openapi/platform.yaml
openapi: 3.0.3
info:
  title: BuildingOS platform foundation API
  version: "0.1.0"
  description: >
    Exact operational routes for the BOS-001 platform foundation milestone.
    Service metadata plus BOS-010 F5a revenue APIs (subscription-service) and F2 building
    application & lifecycle APIs (building-service).
servers:
  - url: http://localhost:8080
    description: Local gateway
paths:
...
  /api/v1/platform/building:
    get:
      summary: Building service metadata (via gateway)
      description: Gateway forwards to building-service's internal /internal/platform/info.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Metadata envelope
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeMetadata" }
        "401":
          description: Missing/expired/wrong-audience/wrong-issuer/bad-signature token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "503":
          description: Downstream building-service unavailable (sanitized)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /api/v1/building-applications:
    post:
      summary: "Start a draft building application (any authenticated user)"
      description: >
        All fields optional while drafting. BOS-010 F2 AP-01/AP-02; building-service.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BuildingApplicationRequest" }
      responses:
        "201":
          description: "Created envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/me/building-applications:
    get:
      summary: "Caller's own applications, newest first"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/building-applications/{applicationId}:
    get:
      summary: "Application detail (applicant or platform admin)"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    put:
      summary: "Replace draft fields (applicant, DRAFT or MORE_INFORMATION_REQUIRED)"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BuildingApplicationRequest" }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "NOT_EDITABLE"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/building-applications/{applicationId}/submit:
    post:
      summary: "Submit (DRAFT/MORE_INFORMATION_REQUIRED to SUBMITTED)"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST or APPLICATION_INCOMPLETE (message lists missing fields)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "INVALID_TRANSITION"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/building-applications/{applicationId}/history:
    get:
      summary: "Audited status transitions (no internal notes)"
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "APPLICATION_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /api/v1/platform/building-applications:
    get:
      summary: "Review queue (drafts never listed)"
      description: >
        meta carries page, size, total.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: integer }
        - name: size
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST (status=DRAFT, size > 100)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /api/v1/platform/buildings/{buildingId}:
    get:
      summary: "Building with memberships"
      security:
        - bearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "BUILDING_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiEnvelope:
      type: object
      required: [success, data, traceId]
      properties:
        success: { type: boolean }
        data: {}
        meta: { type: object }
        traceId: { type: string }
    ApiEnvelopeMetadata:
      type: object
      required: [success, data, meta, traceId]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: object
          required: [service]
          properties:
            service: { type: string, example: auth-service }
        meta: { type: object }
        traceId: { type: string, format: uuid }
    BuildingApplicationRequest:
      type: object
      properties:
        buildingName: { type: string, maxLength: 200 }
        buildingType: { type: string, enum: [RESIDENTIAL, COMMERCIAL, MIXED] }
        address: { type: string, maxLength: 500 }
        area: { type: string, maxLength: 120 }
        district: { type: string, maxLength: 120 }
        postalCode: { type: string, maxLength: 16 }
        totalFloors: { type: integer, minimum: 1 }
        estimatedUnits: { type: integer, minimum: 1 }
        applicantRelationship: { type: string, enum: [OWNER, COMMITTEE_MEMBER, PROPERTY_MANAGER, DEVELOPER, OTHER] }
        relationshipNote: { type: string, maxLength: 200, description: Required at submit when applicantRelationship is OTHER }
        contactName: { type: string, maxLength: 200 }
        contactPhone: { type: string, description: "01XXXXXXXXX, 8801XXXXXXXXX or +8801XXXXXXXXX; stored as 01XXXXXXXXX" }
        contactEmail: { type: string, format: email }
        managementType: { type: string, enum: [SELF_MANAGED, OWNERS_COMMITTEE, MANAGEMENT_COMPANY, DEVELOPER_MANAGED] }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
...
    ApiError:
      type: object
      required: [success, code, message, traceId]
      properties:
        success: { type: boolean, enum: [false] }
        code:
          type: string
          description: Stable error code; each operation lists the codes it may return.
          example: ACCESS_DENIED
        message: { type: string }
        traceId: { type: string, format: uuid }